pyproject.toml
[project]
name = "culicidaelab-server"
version = "0.1.0"
description = "CulicidaeLab server"
readme = "README.md"
requires-python = ">=3.11.0"
dependencies = [
    "aiofiles>=24.1.0",
    "anywidget>=0.9.18",
    "colorcet>=3.1.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "ipyleaflet>=0.19.2",
    "lancedb>=0.22.0",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pyarrow>=20.0.0",
    "pydantic-settings>=2.9.1",
    "pylance>=0.26.1",
    "python-dotenv>=1.1.0",
    "python-i18n[yaml]>=0.3.9",
    "python-multipart>=0.0.20",
    "shapely>=2.1.0",
    "solara[assets]==1.50.1",
    "fastai>=2.7.0, <=2.8.0",
    "culicidaelab>=0.2.2",

]
[dependency-groups]
dev = [
    "black>=24.10.0",
    "docxtpl>=0.20.0",
    "ipykernel>=6.16.2, <=6.29.5",
    "isort>=5.13.2",
    "mypy>=1.13.0",
    "openpyxl>=3.1.5",
    "pre-commit>=4.0.1",
    "pytest-asyncio>=1.0.0",
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "psutil>=7.0.0",
    "tabulate>=0.9.0",
    "types-aiofiles>=24.1.0.20250822",
]

[tool.uv.sources]
culicidaelab-server = { workspace = true }


[tool.setuptools]
packages = ["frontend", "backend"]
